Moved um2nc.drivers.common.DelayedCubePath into um2nc.common
