+++ test/drivers/test_common.py
@@ -303,49 +303,3 @@ def test_input_output_mapping_no_duplicates(input_output, monkeypatch):
     mapping = driver.input_output_mapping
 
     assert mapping == input_output
-
-
-@pytest.mark.parametrize(
-    "p",
-    [
-        "input_path/file.ext",
-        Path("input_path/file.ext"),
-    ]
-)
-def test_DelayedCubePath__init__(p):
-    delayed_path = drivers_common.DelayedCubePath(p)
-    delayed_path.output_path == Path(p)
-
-
-@pytest.mark.parametrize(
-    "varname,expected_error",
-    [
-        # Test if var_name has not been set
-        (None, KeyError),
-        # Test if var_name has been set
-        ("varname", None),
-    ]
-)
-def test_DelayedCubePath__get_var_name(varname, expected_error):
-    cube = Cube([1, 2, 3])
-    if varname:
-        cube.var_name = varname
-
-    if expected_error:
-        with pytest.raises(expected_exception=expected_error):
-            drivers_common.DelayedCubePath._get_var_name(cube)
-    else:
-        field_name = drivers_common.DelayedCubePath._get_var_name(cube)
-        assert field_name == varname
-
-
-def test_DelayedCubePath_resolve_cube():
-    delayed_path = drivers_common.DelayedCubePath("parentdir/filename")
-
-    cube = Cube([1, 2, 3])
-    cube.var_name = "var"
-
-    path = delayed_path.resolve_cube(cube)
-
-    assert isinstance(path, Path)
-    assert path == Path("parentdir/var_filename")
